Fixed select for v1
onboarding-workflow_v1 is now in state when it deploys repos for hardcoded user and does NOT finish onboardng. It does not add user to DAO to let us manually upgrade DAO to v2.0.0 and then run onboarding v2 to add user and finish onboarding.
onboarding-worklow    is onboarding for new users will deploy repos with v2.0.0 for test users worked well
For both versions .env must contain VERSION var (2.0.0 or 1.0.0)